Move error logic from WorkflowDetails to Errors
